webpage/requirements.txt
flask
pandas
numpy
matplotlib
paho-mqtt
plotly
kaleido